Simplified repeat tagging. Add tester
